fix(db): index hashcatMode on the casted int expression (review)
PR review on #179 flagged that the hashcat_mode expression index in
migration 0011 indexed the text value (required_capabilities ->>
'hashcatMode') but the actual hot-path query in services/tasks.ts
buildCapabilityPredicate casts to int:

  (required_capabilities ->> 'hashcatMode')::int = ANY($1::int[])

Postgres expression indexes are only usable when the indexed
expression matches the query expression exactly, so the text index
never served assignNextTask. The intended perf win from P-H3 didn't
land for hashcatMode (the gpu index is fine -- gpu predicate is text
on both sides).

Fix is forward-only migration 0012 that drops + recreates the index
on the casted form. 0011 is updated with a comment noting it is
superseded; not rewritten in place because the wrong-index window is
local-dev only (0011 has not shipped to production yet), and
forward-only migrations are simpler to reason about than rewritten
history. Schema updated to match the casted form so future
db-generate runs stay aligned.

+++ packages/shared/src/db/schema.ts
@@ -455,8 +455,15 @@ export const tasks = pgTable(
     // -- O(unassigned_pending_tasks) per claim at 10-50 claims/sec
     // hits 100-300ms p99 with 100K pending tasks.
     index('tasks_required_capabilities_gpu_idx').on(sql`((required_capabilities ->> 'gpu'))`),
+    // Indexed on the casted int expression so the planner can use it
+    // for the actual hot-path predicate, which does
+    // `(required_capabilities ->> 'hashcatMode')::int = ANY(...)`
+    // (services/tasks.ts buildCapabilityPredicate). Postgres expression
+    // indexes only match when the indexed expression matches the
+    // query expression exactly -- a text-only index would never be
+    // used here despite looking correct on paper.
     index('tasks_required_capabilities_hashcat_mode_idx').on(
-      sql`((required_capabilities ->> 'hashcatMode'))`
+      sql`(((required_capabilities ->> 'hashcatMode'))::int)`
     ),
     // Partial index for the assignNextTask hot path: the planner
     // scans pending/unassigned tasks per claim. Bounding the index
